Revert "feat: build打包构建"
