+++ my-tailc/src/main.ts
@@ -1,8 +1,7 @@
 import { createApp } from "vue";
 import "./style.css";
 import App from "./App.vue";
-import { router } from "./routers";
 
 import "./styles/index.css";
 
-createApp(App).use(router).mount("#app");
+createApp(App).mount("#app");
