--- my-tailc/src/routers/index.ts
@@ -1,11 +1,0 @@
-import { createRouter, createWebHistory } from "vue-router";
-
-export const router = createRouter({
-  history: createWebHistory(),
-  routes: [
-    {
-      path: "/",
-      component: () => import("../views/Home.vue")
-    },
-  ],
-});
