+++ my-tailc/vite.config.ts
@@ -1,21 +1,7 @@
-import { defineConfig, loadEnv } from "vite";
-import vue from "@vitejs/plugin-vue";
-import path from "path";
+import { defineConfig } from 'vite'
+import vue from '@vitejs/plugin-vue'
+
 // https://vitejs.dev/config/
-export default defineConfig(({ mode }) => {
-  const env = loadEnv(mode, __dirname);
-  return {
-    plugins: [vue()],
-    base: env.VITE_MODE === "production" ? "./" : "/",
-    resolve: {
-      alias: {
-        "@": path.resolve(__dirname, "src"),
-      },
-    },
-    build: {
-      outDir: "dist",
-      assetsDir: "assets",
-      sourcemap: env.VITE_MODE === "development" ? true : false,
-    },
-  };
-});
+export default defineConfig({
+  plugins: [vue()]
+})
